+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   [#357](https://github.com/lambda-fairy/maud/pull/357)
 - Support `axum` v0.6 through `axum-core` v0.3
   [#361](https://github.com/lambda-fairy/maud/pull/361)
+- Implement `Default` for `PreEscaped`
+  [#371](https://github.com/lambda-fairy/maud/pull/371)
 
 ## [0.24.0] - 2022-08-12